backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
djangorestframework==3.15.2
openai>=1.0.0,<2.0.0
Pillow>=8.3.2,<10.0.0
requests>=2.26.0,<3.0.0
django-cors-headers
black
stripe
python-dotenv
django-cors-headers